feat: Request model 🧩, database models integration and routes

+++ api/src/controllers/user.controller.ts
@@ -1,5 +1,7 @@
 import { Request, Response } from "express";
-import User from "../models/user";
+import Book from "../models/book";
+import RequestBook from "../models/requestBook";
+import User, { IUser } from "../models/user";
 import { validateToken } from "../services/jwt.service";
 
 export type userDataType = {
@@ -16,23 +18,35 @@ export const getAllUsers = async (_: Request, res: Response) => {
   }
 };
 
-export const getUser = async (req: Request, res: Response) => {
-  const token = req.headers["x-access-token"] as string;
-  if (!token) {
-    return res.status(401).send({ message: "No token provided" });
-  }
+export const getAllAdmins = async (_: Request, res: Response) => {
   try {
-    const user = validateToken(token);
-    res.send(user);
+    const users = await User.find({ role: "admin" });
+    if (!users) {
+      return res.status(404).send({ message: "No admins found" });
+    }
+    res.send(users);
   } catch (error) {
-    res.status(400).send({ message: "Invalid token" });
+    res.status(400).send(error);
   }
+};
+
+export const getAllStudents = async (_: Request, res: Response) => {
+  try {
+    const users = await User.find({ role: "user" });
+    if (!users) {
+      return res.status(404).send({ message: "No users found" });
+    }
+    res.send(users);
+  } catch (error) {}
 };
 
 export const getUserById = async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
     const user = await User.findById(id);
+    if (!user) {
+      return res.status(404).send({ message: "User not found" });
+    }
     return res.status(202).json(user);
   } catch (error) {
     res.status(400).send({ message: "Invalid token" });
@@ -41,9 +55,20 @@ export const getUserById = async (req: Request, res: Response) => {
 
 export const updateUserById = async (req: Request, res: Response) => {
   const { id } = req.params;
-  const updates = req.body;
+  const updates = req.body as IUser;
   try {
-    const user = await User.findByIdAndUpdate(id, updates, { new: true });
+    const { name, email, role, password, borrowedBooks } = updates;
+    if (password)
+      return res.status(400).send({ message: "Password cannot be updated" });
+    if (borrowedBooks)
+      return res
+        .status(400)
+        .send({ message: "Borrowed books cannot be updated" });
+    const user = await User.findByIdAndUpdate(
+      id,
+      { name, email, role },
+      { new: true }
+    );
     res.status(200).send({ message: `User sucessfuly update`, user });
   } catch (error) {
     res.status(400).send(error);
@@ -53,7 +78,24 @@ export const updateUserById = async (req: Request, res: Response) => {
 export const deleteUserById = async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
-    const user = await User.findByIdAndDelete(id);
+    const user = (await User.findById(id)) as IUser;
+    if (!user) return res.status(404).send({ message: "User not found" });
+    user.borrowedBooks.forEach(async (book) => {
+      if (!book.returnDate) {
+        Book.findOneAndUpdate(
+          { _id: book.bookId },
+          { returnDate: new Date() },
+          { new: true }
+        );
+      }
+    });
+
+    await RequestBook.updateMany(
+      { userId: { $eq: id } },
+      { status: "deleted" }
+    );
+
+    await User.findByIdAndDelete(id);
     res.status(200).send({ message: `User sucessfuly deleted`, user });
   } catch (error) {
     res.status(400).send(error);
@@ -64,6 +106,8 @@ export const deteteAllUsers = async (_: Request, res: Response) => {
   try {
     const user = await User.deleteMany();
     res.status(200).send({ message: `All users sucessfuly deleted`, user });
+    await RequestBook.updateMany({}, { status: "deleted" });
+    await Book.updateMany({}, { returnDate: new Date() });
   } catch (error) {
     res.status(400).send(error);
   }
